Revert build to pre-c++11 so CentOS 6.6 will compile without upgrading gcc.

+++ include/api/seabreezeapi/SpectrometerFeatureAdapter.h
@@ -64,7 +64,7 @@ class SpectrometerFeatureAdapter: public FeatureAdapterTemplate<OOISpectrometerF
 	int getOpticalDarkPixelIndices(int *errorCode, int *indices, int length);
 	int getActivePixelCount(int *errorCode);
 	int getActivePixelIndices(int *errorCode, int *indices, int length);
-	unsigned int getNumberOfPixels(int *errorCode);
+	unsigned short getNumberOfPixels(int *errorCode);
 	void setIntegrationTimeMicros(int *errorCode,
 		unsigned long integrationTimeMicros);
 	long getMinimumIntegrationTimeMicros(int *errorCode);
