+++ include/vendors/OceanOptics/features/spectrometer/OOISpectrometerFeature.h
@@ -87,7 +87,7 @@ class OOISpectrometerFeature: public FeatureImpl,
 	virtual long getIntegrationTimeMaximum() const;
 	virtual long getIntegrationTimeIncrement() const;
 
-	virtual int getNumberOfPixels() const;
+	virtual unsigned short getNumberOfPixels() const;
 	virtual int getMaximumIntensity() const;
 
 	/* Overriding from Feature */
@@ -98,8 +98,8 @@ class OOISpectrometerFeature: public FeatureImpl,
 	IntrospectionFeature *myIntrospection;
 
 	/* Detector details */
-	uint16_t numberOfPixels;
-	uint32_t maxIntensity;
+	unsigned short numberOfPixels;
+	unsigned int maxIntensity;
 
 	/* Integration time parameters (measured in microseconds) */
 	long integrationTimeMinimum;
